move sample folder to tests and refactor paths accordingly
